M1d: end-to-end network loopback (capture -> encode -> TCP -> decode -> render)
Turn the host/client skeletons into working binaries that stream the host's
display to a client window over TCP, completing milestone M1.

- host: ScreenCaptureKit capture -> VideoToolbox H.264 encode (zero-copy
  IOSurface bridge) -> TcpListener streaming StreamStart(SPS/PPS) + Frames
- client: rebuild CMFormatDescription + reassemble CMSampleBuffer from the
  wire, decode (NV12), PixelTransfer NV12->BGRA, render via wgpu (reuses the
  loopback_viewer path); adds its render/codec deps + a Swift-runtime build.rs
- protocol: add portable is_keyframe(codec, data) with unit tests
- examples/reassembly_probe: self-verifiable encode -> wire -> decode roundtrip

Verified end-to-end over localhost (host streamed 2240x1260 H.264; client
reconstructed and decoded continuously, no errors). clippy --all-targets and
tests are green.

+++ crates/host/examples/loopback_viewer.rs
@@ -92,9 +92,9 @@ fn start_capture(shared: Shared) -> Option<SCStream> {
             let Ok(guard) = buffer.lock(CVPixelBufferLockFlags::READ_ONLY) else {
                 return;
             };
-            let w = guard.width() as usize;
-            let h = guard.height() as usize;
-            let stride = guard.bytes_per_row() as usize;
+            let w = guard.width();
+            let h = guard.height();
+            let stride = guard.bytes_per_row();
             let row = w * 4;
             let src = guard.as_slice();
             if w == 0 || h == 0 || stride < row || src.len() < stride * h {
